Rename to objectWithOnly and make it non recursive
