+++ MATCHERS.md
@@ -4,13 +4,13 @@
 - [boolean](#boolean)
 - [duration](#duration)
 - [enum](#enum)
-- [exactObject](#exactObject)
 - [func](#func)
 - [hashmap](#hashmap)
 - [integer](#integer)
 - [isoDate](#isoDate)
 - [number](#number)
 - [object](#object)
+- [objectWithOnly](#objectWithOnly)
 - [regex](#regex)
 - [string](#string)
 - [url](#url)
@@ -70,36 +70,6 @@ state: s.enum({values: ['NSW', 'VIC'], name:'state'})
 // displays "should be a valid state (NSW,VIC)"
 state: s.enum({values: ['NSW', 'VIC'], name:'state', verbose: true})
 ```
-
-## exactObject
-
-Match an exact object.
-This matcher will not error if any optional properties are left out.
-This matcher will error if any extra properties passed in.
-
-Each property must have a corresponding matcher, and they will be called recursively.
-
-```js
-// match exact object
-person: 'exactObject'
-person: s.exactObject()
-
-// match an object with given properties
-person: s.exactObject({
-  name: 'string'
-  age: 'number'
-})
-
-// match a nested object
-person: s.exactObject({
-  name: 'string'
-  address: {
-    city: 'string'
-    postcode: 'number'
-  }
-})
-```
-
 
 ## func
 
@@ -202,6 +172,31 @@ person: {
     postcode: 'number'
   }
 }
+```
+
+## objectWithOnly
+
+Match an exact object.
+This matcher will not error if any optional properties are left out.
+This matcher will error if any extra properties are passed in.
+
+
+```js
+
+// match an object with given properties
+person: s.objectWithOnly({
+  name: 'string'
+  age: 'number'
+})
+
+// match a nested object
+person: s.objectWithOnly({
+  name: 'string'
+  address: s.objectWithOnly({
+    city: 'string'
+    postcode: 'number'
+  })
+})
 ```
 
 ## regex
