+++ README.md
@@ -81,13 +81,13 @@ Some of the most common built-in matchers are
 - `s.boolean()`
 - `s.duration(min, max)`
 - `s.enum(name, values, verbose)`
-- `s.exactObject(fields)`
 - `s.func(arity)`
 - `s.hashmap(keys, values)`
 - `s.integer(min, max)`
 - `s.isoDate()`
 - `s.number(min, max)`
 - `s.object(fields)`
+- `s.objectWithOnly(fields)`
 - `s.regex(reg)`
 - `s.string()`
 - `s.url()`
