+++ lib/matchers/objectWithOnly.js
@@ -20,11 +20,6 @@ module.exports = function(spec) {
             value: val[key],
             message: 'should not exist'
           });
-        } else if (typeof spec[key] === 'object') {
-          var subpath = path ? (path + '.' + key) : key;
-          var matcher = s.exactObject(spec[key]);
-          var err = matcher(subpath, val[key]);
-          if (err) errors.push(err)
         }
       }
       return _.flatten(errors);
